+++ lib/easytest/matcher/nil.rb
@@ -1,10 +1,6 @@
 module Easytest
   module Matcher
     class Nil < Base
-      def initialize(actual:, negate:)
-        super(actual: actual, expected: nil, negate: negate)
-      end
-
       def match?
         actual.nil?
       end