+++ lib/easytest/matcher/true.rb
@@ -1,10 +1,6 @@
 module Easytest
   module Matcher
     class True < Base
-      def initialize(actual:, negate:)
-        super(actual: actual, expected: true, negate: negate)
-      end
-
       def match?
         actual == expected
       end